add routing, rearrange components
