+++ src/Components/Christmas.css
@@ -1,5 +1,5 @@
 
-.App-header {
+.App-header.christmas {
   background-color: #d42426;
 }
